adding more spells, close to done (missing laser + arcspread)
