+++ enemies/bosses/stompy.lua
@@ -88,8 +88,6 @@ fns['init_enemy'] = function(self)
       group = main.current.main,
       team = "enemy",
       spell_duration = 10,
-      x = self.x,
-      y = self.y,
       num_shots = 3,
       shot_interval = 0.7,
       dmg = 30,
