--- enemies/enemy_attacks/enemy_attack_includes.lua
@@ -1,3 +1,0 @@
-require "enemies/enemy_attacks/stomp"
-require "enemies/enemy_attacks/mortar"
-require "enemies/enemy_attacks/avalanche"
