+++ enemies/enemy_critter.lua
@@ -142,83 +142,3 @@ function EnemyCritter:on_trigger_enter(other, contact)
     other:hit(self.dmg, nil, nil, true)
   end]]--
 end
-
-Summon = Object:extend()
-Summon:implement(GameObject)
-Summon:implement(Physics)
-function Summon:init(args)
-  self:init_game_object(args)
-  self.attack_sensor = Circle(self.x, self.y, self.rs)
-  self.currentTime = 0
-
-  self.state = "charging"
-
-  self.parent.state = 'frozen'
-  
-  self.color_transparent = self.color:clone()
-  self.color_transparent.a = 0.4
-
-  self.summonTime = self.castTime or 3
-  illusion1:play{pitch = random:float(0.8, 1.2), volume = 0.5}
-  self.t:after(self.summonTime, function() self:spawn() end)
-
-end
-
-function Summon:update(dt)
-  if self.parent and self.parent.dead then self.dead = true; return end
-  self:update_game_object(dt)
-  self.x = self.parent.x
-  self.y = self.parent.y
-  self.currentTime = self.currentTime + dt
-end
-
-function Summon:spawn()
-  illusion1:play{pitch = random:float(0.8, 1.2), volume = 0.5}
-  spawn1:play{pitch = random:float(0.8, 1.2), volume = 0.15}
-  if self.parent.summons and self.parent.maxSummons and self.parent.summons < self.parent.maxSummons then
-    if self.type == 'enemy_critter' then
-      for i = 1, self.amount do
-        local offset = SpawnGlobals.spawn_offsets[i % #SpawnGlobals.spawn_offsets]
-        local x, y = self.x + offset.x, self.y + offset.y
-        if Can_Spawn(2, {x = x, y = y}) then
-            self.parent.summons = self.parent.summons + 1
-            EnemyCritter{group = main.current.main, x = x, y = y, color = grey[0], r = random:float(0, 2*math.pi), 
-            v = 10, parent = self.parent}
-        end
-      end
-    else
-      for i = 1, self.amount do
-        local offset = SpawnGlobals.spawn_offsets[i % #SpawnGlobals.spawn_offsets]
-        local x, y = self.x + offset.x, self.y + offset.y
-        if Can_Spawn(6, {x = x, y = y}) then
-            self.parent.summons = self.parent.summons + 1
-            Enemy{type = self.type, group = main.current.main, x = self.x + offset.x, y = self.y + offset.y, level = self.level, parent = self.parent}
-        end
-      end
-    end
-  end
-  self:recover()
-end
-
-function Summon:recover()
-  self.state = 'recovering'
-  if self and self.parent then self.parent.state = 'normal' end
-  if self.suicide and self.parent then
-    self.parent:die()
-    self.dead = true
-  end
-  if self then self.dead = true end
-end
-
-function Summon:draw()
-  if self.hidden then return end
-
-  if self.state == 'charging' then
-    graphics.push(self.x, self.y, self.r + (self.vr or 0), self.spring.x, self.spring.x)
-      -- graphics.circle(self.x, self.y, self.shape.rs + random:float(-1, 1), self.color, 2)
-      graphics.circle(self.x, self.y, self.attack_sensor.rs * math.min(self.currentTime / self.summonTime, 1) , self.color_transparent)
-      graphics.circle(self.x, self.y, self.attack_sensor.rs, self.color, 1)
-    graphics.pop()
-  end
-
-end
